% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_dataset_summary.R
\name{print.training_dataset_summary}
\alias{print.training_dataset_summary}
\title{Print method for training_dataset_summary}
\usage{
\method{print}{training_dataset_summary}(x, ...)
}
\arguments{
\item{x}{A training_dataset_summary object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for training_dataset_summary
}
\seealso{
Other audio:
\code{\link[=analyze_call_properties]{analyze_call_properties()}},
\code{\link[=biggest_files]{biggest_files()}},
\code{\link[=convert_to_wav]{convert_to_wav()}},
\code{\link[=find_duplicate_wavs]{find_duplicate_wavs()}},
\code{\link[=training_dataset_summary]{training_dataset_summary()}}
}
\concept{audio}
