% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{skip_if_api_unavailable}
\alias{skip_if_api_unavailable}
\title{Skip test if API is unreachable}
\usage{
skip_if_api_unavailable(api_type)
}
\arguments{
\item{api_type}{Character. "inat" or "ala".}
}
\description{
Helper for testthat to elegantly skip integration tests when APIs are down.
}
\keyword{internal}
