% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{is_api_reachable}
\alias{is_api_reachable}
\title{Check if a remote API is reachable}
\usage{
is_api_reachable(api_type = c("inat", "ala"), timeout_secs = 2)
}
\arguments{
\item{api_type}{Character. Either "inat" or "ala".}

\item{timeout_secs}{Numeric. Seconds to wait before timing out. Default is 2.}
}
\value{
Logical. TRUE if reachable, FALSE otherwise.
}
\description{
Check if a remote API is reachable
}
\keyword{internal}
