% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_dataset_summary.R
\name{print.training_dataset_summary}
\alias{print.training_dataset_summary}
\title{Print method for training_dataset_summary}
\usage{
\method{print}{training_dataset_summary}(x, ...)
}
\arguments{
\item{x}{A training_dataset_summary object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for training_dataset_summary
}
